refactor: bonuses reduces
