+++ ts/features/bonus/bonusVacanze/screens/AvailableBonusScreen.tsx
@@ -48,14 +48,6 @@ import {
   AvailableBonusItemState
 } from "../components/AvailableBonusItem";
 import { navigateToBonusRequestInformation } from "../navigation/action";
-import { loadAvailableBonuses } from "../store/actions/bonusVacanze";
-import {
-  experimentalAndVisibleBonus,
-  isAvailableBonusLoadingSelector,
-  isAvailableBonusNoneErrorSelector,
-  serviceFromAvailableBonusSelector,
-  supportedAvailableBonusSelector
-} from "../store/reducers/availableBonusesTypes";
 import {
   ID_BONUS_VACANZE_TYPE,
   ID_BPD_TYPE,
@@ -71,6 +63,14 @@ import {
   showServiceDetails
 } from "../../../../store/actions/services";
 import { IOStyles } from "../../../../components/core/variables/IOStyles";
+import {
+  experimentalAndVisibleBonus,
+  isAvailableBonusLoadingSelector,
+  isAvailableBonusNoneErrorSelector,
+  serviceFromAvailableBonusSelector,
+  supportedAvailableBonusSelector
+} from "../../common/store/selectors";
+import { loadAvailableBonuses } from "../../common/store/actions/availableBonusesTypes";
 
 export type Props = ReturnType<typeof mapStateToProps> &
   ReturnType<typeof mapDispatchToProps>;
