+++ ts/features/bonus/bonusVacanze/store/reducers/index.ts
@@ -1,31 +1,15 @@
 import { combineReducers } from "redux";
 import { Action } from "../../../../../store/actions/types";
-import bpdReducer, { BpdState } from "../../../bpd/store/reducers";
-import cgnReducer, { CgnState } from "../../../cgn/store/reducers";
-import svReducer, { SvState } from "../../../siciliaVola/store/reducers";
-import cdcReducer, { CdcState } from "../../../cdc/store/reducers";
 import bonusVacanzeActivationReducer, { ActivationState } from "./activation";
 import allActiveReducer, { AllActiveState } from "./allActive";
-import availableBonusesReducer, {
-  AvailableBonusTypesState
-} from "./availableBonusesTypes";
 import eligibilityReducer, { EligibilityState } from "./eligibility";
 
 // type alias
-type BonusVacanzeState = {
+export type BonusVacanzeState = {
   activation: ActivationState;
   eligibility: EligibilityState;
   allActive: AllActiveState;
 };
-
-export type BonusState = Readonly<{
-  availableBonusTypes: AvailableBonusTypesState;
-  bonusVacanze: BonusVacanzeState;
-  bpd: BpdState;
-  cgn: CgnState;
-  sv: SvState;
-  cdc: CdcState;
-}>;
 
 const bonusVacanzeReducer = combineReducers<BonusVacanzeState, Action>({
   activation: bonusVacanzeActivationReducer,
@@ -33,13 +17,4 @@ const bonusVacanzeReducer = combineReducers<BonusVacanzeState, Action>({
   allActive: allActiveReducer
 });
 
-const bonusReducer = combineReducers<BonusState, Action>({
-  availableBonusTypes: availableBonusesReducer,
-  bonusVacanze: bonusVacanzeReducer,
-  bpd: bpdReducer,
-  cgn: cgnReducer,
-  sv: svReducer,
-  cdc: cdcReducer
-});
-
-export default bonusReducer;
+export default bonusVacanzeReducer;
