+++ ts/features/bonus/bpd/screens/onboarding/BpdCTAStartOnboardingScreen.tsx
@@ -10,11 +10,11 @@ import { bpdRemoteConfigSelector } from "../../../../../store/reducers/backendSt
 import { GlobalState } from "../../../../../store/reducers/types";
 import { useActionOnFocus } from "../../../../../utils/hooks/useOnFocus";
 import { LoadingErrorComponent } from "../../../bonusVacanze/components/loadingErrorScreen/LoadingErrorComponent";
-import { loadAvailableBonuses } from "../../../bonusVacanze/store/actions/bonusVacanze";
+import { loadAvailableBonuses } from "../../../common/store/actions/availableBonusesTypes";
 import {
   isAvailableBonusErrorSelector,
   supportedAvailableBonusSelector
-} from "../../../bonusVacanze/store/reducers/availableBonusesTypes";
+} from "../../../common/store/selectors";
 import { bpdOnboardingStart } from "../../store/actions/onboarding";
 
 export type Props = ReturnType<typeof mapDispatchToProps> &
