+++ ts/features/bonus/bpd/screens/onboarding/BpdInformationScreen.tsx
@@ -6,7 +6,6 @@ import { Dispatch } from "../../../../../store/actions/types";
 import { GlobalState } from "../../../../../store/reducers/types";
 import { emptyContextualHelp } from "../../../../../utils/emptyContextualHelp";
 import { actionWithAlert } from "../../../common/components/alert/ActionWithAlert";
-import { availableBonusTypesSelectorFromId } from "../../../bonusVacanze/store/reducers/availableBonusesTypes";
 import { ID_BPD_TYPE } from "../../../bonusVacanze/utils/bonus";
 import BonusInformationComponent from "../../../common/components/BonusInformationComponent";
 import {
@@ -14,6 +13,7 @@ import {
   bpdUserActivate
 } from "../../store/actions/onboarding";
 import { bpdEnabledSelector } from "../../store/reducers/details/activation";
+import { availableBonusTypesSelectorFromId } from "../../../common/store/selectors";
 
 export type Props = ReturnType<typeof mapDispatchToProps> &
   ReturnType<typeof mapStateToProps>;
