+++ ts/features/bonus/cgn/components/CgnServiceCTA.tsx
@@ -10,7 +10,6 @@ import { useIODispatch, useIOSelector } from "../../../../store/hooks";
 import { servicePreferenceSelector } from "../../../../store/reducers/entities/services/servicePreference";
 import { isServicePreferenceResponseSuccess } from "../../../../types/services/ServicePreferenceResponse";
 import { ServiceId } from "../../../../../definitions/backend/ServiceId";
-import { loadAvailableBonuses } from "../../bonusVacanze/store/actions/bonusVacanze";
 import { cgnActivationStart } from "../store/actions/activation";
 import { cgnUnsubscribe } from "../store/actions/unsubscribe";
 import { fold, isLoading } from "../../bpd/model/RemoteValue";
@@ -18,6 +17,7 @@ import { showToast } from "../../../../utils/showToast";
 import { cgnUnsubscribeSelector } from "../store/reducers/unsubscribe";
 import { loadServicePreference } from "../../../../store/actions/services/servicePreference";
 import ActivityIndicator from "../../../../components/ui/ActivityIndicator";
+import { loadAvailableBonuses } from "../../common/store/actions/availableBonusesTypes";
 
 type Props = {
   serviceId: ServiceId;
