+++ ts/features/wallet/component/__test__/FeaturedCardCarousel.test.tsx
@@ -4,7 +4,6 @@ import FeaturedCardCarousel from "../card/FeaturedCardCarousel";
 import { appReducer } from "../../../../store/reducers";
 import { bpdLoadActivationStatus } from "../../../bonus/bpd/store/actions/details";
 import { GlobalState } from "../../../../store/reducers/types";
-import { loadAvailableBonuses } from "../../../bonus/bonusVacanze/store/actions/bonusVacanze";
 import { availableBonuses } from "../../../bonus/bonusVacanze/__mock__/availableBonuses";
 import * as bonus from "../../../bonus/bonusVacanze/utils/bonus";
 import {
@@ -16,6 +15,7 @@ import { BonusVisibilityEnum } from "../../../../../definitions/content/BonusVis
 import * as cgnDetailSelectors from "../../../bonus/cgn/store/reducers/details";
 import { renderScreenFakeNavRedux } from "../../../../utils/testWrapper";
 import { MESSAGES_ROUTES } from "../../../messages/navigation/routes";
+import { loadAvailableBonuses } from "../../../bonus/common/store/actions/availableBonusesTypes";
 
 jest.mock("react-native-share", () => jest.fn());
 describe("FeaturedCardCarousel", () => {
